extracted getNodeProp util, add unit test
