+++ __tests__/utils-test.js
@@ -1,0 +1,24 @@
+const getNodeProperty = require('../src/utils');
+
+describe('Unit | utils | getNodeProperty', () => {
+  it('returns property as expected', () => {
+    const mockNode = {
+      system: {
+        settings: {
+          volume: {
+            level: 11,
+          },
+        },
+      },
+    };
+
+    expect(getNodeProperty(mockNode, 'system.settings.volume.level')).toBe(11);
+    expect(getNodeProperty(mockNode.system.settings, 'volume.level')).toBe(11);
+    expect(getNodeProperty(mockNode, 'unknown.leaf.property')).toBe(undefined);
+  });
+
+  it('returns early if no node is provided', () => {
+    let emptyNode;
+    expect(getNodeProperty(emptyNode, 'volume.level')).toBe(undefined);
+  });
+});
