+++ package.json
@@ -63,7 +63,7 @@
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-prettier": "^3.4.0",
     "husky": "^1.2.0",
-    "jest": "^23.6.0",
+    "jest": "^27.0.6",
     "lint-staged": "^8.1.0",
     "release-it": "^14.2.1",
     "release-it-lerna-changelog": "^3.1.0"
