+++ src/index.js
@@ -1,4 +1,6 @@
 const path = require('path');
+// eslint-disable-next-line node/no-unpublished-require
+const getNodeProperty = require('../src/utils.js');
 
 /**
  * Checks if the call expression matches a test setup call pattern.
@@ -186,38 +188,6 @@ function hasMetadataDeclaration({ node }, t) {
     node.expression.left.object.name === 'testMetadata' &&
     node.expression.left.property.name === 'filePath'
   );
-}
-
-/**
- * Utility to get a property from a given path
- * @param {object} node
- * @param {string} path
- * @returns property value
- */
-function getNodeProperty(node, path) {
-  if (!node) {
-    return;
-  }
-
-  let parts;
-  if (typeof path === 'string') {
-    parts = path.split('.');
-  } else {
-    parts = path;
-  }
-
-  if (parts.length === 1) {
-    return node[path];
-  }
-
-  let property = node[parts[0]];
-
-  if (property && parts.length > 1) {
-    parts.shift();
-    return getNodeProperty(property, parts);
-  }
-
-  return property;
 }
 
 /**
