+++ src/utils.js
@@ -1,0 +1,33 @@
+/**
+ * Utility to get a property from a given path
+ * @param {object} node
+ * @param {string} path
+ * @returns property value
+ */
+function getNodeProperty(node, path) {
+  if (!node) {
+    return;
+  }
+
+  let parts;
+  if (typeof path === 'string') {
+    parts = path.split('.');
+  } else {
+    parts = path;
+  }
+
+  if (parts.length === 1) {
+    return node[path];
+  }
+
+  let property = node[parts[0]];
+
+  if (property && parts.length > 1) {
+    parts.shift();
+    return getNodeProperty(property, parts);
+  }
+
+  return property;
+}
+
+module.exports = getNodeProperty;
